Extract language into its dedicated class
